ENH: trained model for all site-wide comments in eda.ipynb
Renmaed the eda notebook to indicate it is for the site-wide survey. I also trained a ham-spam model that looks at all four of the comment fields at once. This required hand-labeling approximately 10k survey responses. Also, instead of using word embeddings, which are expensive and large to persist, I used tf-idf, which trains considerably faster and doesn't take up a lot of disk space.